move storage to new file
